[chromeperf] Pinpoint use the patchset before rebase if the CL has landed
When using a landed CL as patch in Pinpoint, the change will be checked out
instead of cherry-picked. This caused unexpected behavior to developers.

On Gerrit, the last patchset should be a rebase from the HEAD at that time.
This CL allow Pinpoint to use the patchset before rebase so that this change
will be cherry-picked correctly.

Bug: 460800062
Change-Id: I6281d97cf5bc7a1d93bdbbb3fc9326555e93fc68
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7174076
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>