Surface websocket connection loss as a crash
When converting an inspector websocket error into a Telemetry error,
treat WebSocketConnectionClosedException as an indication that the
target crashed and surface it as such. Previously, all cases of
WebSocketException would be translated into a TimeoutError.

This is a speculative fix for some flakes in tests which purposely crash
the renderer process and expect a DevtoolsTargetCrashException as a
result. These tests are flakily getting TimeoutErrors even though a
renderer crash minidump is generated, indicating that the the target did
actually crash.

Bug: 462990046
Change-Id: Id94227c9b233f645fcae166460ea34a32191a4c2
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7254657
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>