[chromeperf] Pinpoint cancel all swarming builds if a try job fails
In a try job, if one branch has the build failed, the Pinpoint job will
fail. However, the other build of the other branch will continue and
thus become a waste of builder resource.

This CL tries to cancel the builds when a try job fails.

To limit the scope, it only handles:
 - when the job type is try job.
 - only cancel the builds, not the test runs.
 - only when the Pinpoint job fails, not when it is canceled.

Change-Id: If465a3cf365c73bd36d0b8dbee976e9a45c9f5c4
Bug: 445817923
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7282107
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>